change_localization/requirements.txt
matplotlib==3.4.3
numpy==1.21.2
Pillow=9.4.0
torch==1.12.0
torchvision=0.13.0
tqdm==4.62.3                                            

